+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add macOS support for Swiftly toolchain management ([#1673](https://github.com/swiftlang/vscode-swift/pull/1673))
 - Show revision hash or local/editing keyword in project panel dependency descriptions ([#1667](https://github.com/swiftlang/vscode-swift/pull/1667))
 
 ## 2.6.2 - 2025-07-02